fix Problem cast double to decimal

+++ SpicyLand/Models/Panino.cs
@@ -7,7 +7,7 @@ namespace SpicyLand.Models
         public Guid PaninoID { get; set; }
         public string Nome { get; set; } = "";
         public string Descrizione { get; set; } = "";
-        public double Prezzo { get; set; }
+        public decimal Prezzo { get; set; }
         public DateTime DataCreazione { get; set; }
         public IFormFile Immagine { get; set; }
         public string PathImage { get; set; } = "";
